requirements.txt
alembic==1.16.4
amqp==5.3.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Admin==1.6.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
limits==5.5.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ngrok==1.5.1
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.4.0
redis-server==6.0.9
requests==2.32.5
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
stripe==12.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
